move util fn to utils
